mrv-ai-preview-back/pyproject.toml
[project]
name = "workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain-google-genai>=2.0.10",
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.1,<5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.0",
    "pydantic[email]>=2.0.0",
    "email-validator>=2.0.0",
]
